completed lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,8 +4,8 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient= (double)top/bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
@@ -19,7 +19,6 @@ public class DebugExercise1 {
         int b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
-
         int t3 = 3;
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
